backend/requirements.txt
fastapi==0.116.1
requests==2.32.5
google-cloud-storage==3.3.0
uvicorn==0.35.0
dotenv==0.9.9